### 2024-03-22 3ac22b5
Create ci.yml
Add coverage badge
