+++ tox.ini
@@ -1,0 +1,22 @@
+[testenv]
+package = wheel
+
+[coverage:run]
+branch = True
+
+[coverage:report]
+; Regexes for lines to exclude from consideration
+exclude_also =
+
+    ; Don’t complain if tests don’t hit defensive assertion code:
+    raise AssertionError
+    raise NotImplementedError
+
+    ; Don’t complain if non-runnable code isn’t run:
+    if __name__ == .__main__.:
+
+ignore_errors = True
+
+[coverage:html]
+directory = coverage_html_report
+
