+++ .github/workflows/ci.yml
@@ -58,7 +58,7 @@ jobs:
       - name: "Set up Python"
         uses: "actions/setup-python@v4"
         with:
-          python-version: "3.11"
+          python-version: "3.10"
 
       - name: "Install dependencies"
         run: |
