+++ .github/workflows/ci.yml
@@ -21,7 +21,7 @@ jobs:
           - macos-latest
           - windows-latest
         python-version:
-          - "${{ matrix.python-version }}"
+          - "3.10"
 
     steps:
       - name: "Check out the repo"
@@ -58,7 +58,7 @@ jobs:
       - name: "Set up Python"
         uses: "actions/setup-python@v4"
         with:
-          python-version: "3.10"
+          python-version: "{{ matrix.python-version }}"
 
       - name: "Install dependencies"
         run: |
