+++ .github/workflows/ci.yml
@@ -43,7 +43,7 @@ jobs:
       - name: "Upload coverage data"
         uses: actions/upload-artifact@v3
         with:
-          name: "coverage-data"
+          name: "coverage-data-${{ matrix.python-version }}"
           path: .coverage.*
 
   coverage:
@@ -66,7 +66,8 @@ jobs:
       - name: "Download coverage data"
         uses: actions/download-artifact@v3
         with:
-          name: "coverage-data"
+          name: "coverage-data-${{ matrix.python-version }}"
+          merge-multiple: true
 
       - name: "Combine coverage and fail it it’s under 100 %"
         run: |
