+++ tox.ini
@@ -4,6 +4,18 @@ package = wheel
 [coverage:run]
 branch = True
 
+[testenv]
+commands =
+    python -m coverage run -p -m pytest
+
+[testenv:coverage]
+basepython = python3.10
+commands =
+    python -m coverage combine
+    python -m coverage report -m --skip-covered
+    python -m coverage json
+
+
 [coverage:report]
 ; Regexes for lines to exclude from consideration
 exclude_also =
