+++ tox.ini
@@ -1,10 +1,9 @@
-[testenv]
-package = wheel
 
 [coverage:run]
 branch = True
 
 [testenv]
+package = wheel
 commands =
     python -m coverage run -p -m pytest
 
