+++ tox.ini
@@ -1,9 +1,7 @@
-
-[coverage:run]
-branch = True
+[tox]
+envlist = py310,coverage
 
 [testenv]
-package = wheel
 commands =
     python -m coverage run -p -m pytest
 
@@ -14,20 +12,7 @@ commands =
     python -m coverage report -m --skip-covered
     python -m coverage json
 
-
-[coverage:report]
-; Regexes for lines to exclude from consideration
-exclude_also =
-
-    ; Don’t complain if tests don’t hit defensive assertion code:
-    raise AssertionError
-    raise NotImplementedError
-
-    ; Don’t complain if non-runnable code isn’t run:
-    if __name__ == .__main__.:
-
-ignore_errors = True
-
-[coverage:html]
-directory = coverage_html_report
+[gh-actions]
+python =
+    3.10: py310
 
