+++ .github/workflows/ci.yml
@@ -1,14 +1,12 @@
-# Run tests
+# Run CI
 
-name: "Test"
+name: "CI"
 
 on:
   push:
+    branches: [main, tests]
   pull_request:
-
-defaults:
-  run:
-    shell: bash
+    branches: [main]
 
 jobs:
   tests:
@@ -27,10 +25,10 @@ jobs:
 
     steps:
       - name: "Check out the repo"
-        uses: "actions/checkout@v2"
+        uses: "actions/checkout@v3"
 
       - name: "Set up Python"
-        uses: "actions/setup-python@v2"
+        uses: "actions/setup-python@v4"
         with:
           python-version: "${{ matrix.python-version }}"
 
@@ -45,37 +43,49 @@ jobs:
       - name: "Upload coverage data"
         uses: actions/upload-artifact@v3
         with:
-          name: covdata
+          name: "coverage-data-${{ matrix.python-version }}"
           path: .coverage.*
 
   coverage:
-    name: Coverage
+    name: "Combine and check coverage"
     needs: tests
     runs-on: ubuntu-latest
     steps:
       - name: "Check out the repo"
-        uses: "actions/checkout@v2"
+        uses: "actions/checkout@v3"
 
       - name: "Set up Python"
-        uses: "actions/setup-python@v2"
+        uses: "actions/setup-python@v4"
         with:
-          python-version: "3.10"
+          python-version: "{{ matrix.python-version }}"
 
       - name: "Install dependencies"
         run: |
-          python -m pip install tox tox-gh-actions
+          python -m pip install --upgrade coverage[toml]
 
       - name: "Download coverage data"
         uses: actions/download-artifact@v3
         with:
-          name: covdata
+          name: "coverage-data-${{ matrix.python-version }}"
+          merge-multiple: true
 
-      - name: "Combine"
+      - name: "Combine coverage and fail it it’s under 100 %"
         run: |
-          python -m tox -e coverage
-          export TOTAL=$(python -c "import json;print(json.load(open('coverage.json'))['totals']['percent_covered_display'])")
-          echo "total=$TOTAL" >> $GITHUB_ENV
-          echo "### Total coverage: ${TOTAL}%" >> $GITHUB_STEP_SUMMARY
+          python -m coverage combine
+          python -m coverage html --skip-covered --skip-empty
+
+          # Report and write to summary.
+          python -m coverage report | sed 's/^/    /' >> $GITHUB_STEP_SUMMARY
+
+          # Report again and fail if under 100%.
+          python -Im coverage report --fail-under=100
+
+      - name: "Upload HTML report if check failed"
+        uses: actions/upload-artifact@v3
+        with:
+          name: html-report
+          path: htmlcov
+        if: ${{ failure() }}
 
       - name: "Create badge"
         uses: schneegans/dynamic-badges-action@v1.6.0
