+++ tox.ini
@@ -3,7 +3,7 @@ envlist = py310,coverage
 
 [testenv]
 commands =
-    python -m coverage run -p -m pytest
+    python -m coverage run -p -m pytest {posargs}
 
 [testenv:coverage]
 basepython = python3.10
