+++ tox.ini
@@ -1,16 +1,24 @@
 [tox]
-envlist = py310,coverage
+envlist = py310-{test},coverage
 
 [testenv]
+package = wheel
+wheel_build_env = .pkg
+extras =
+    test: test
 commands =
     python -m coverage run -p -m pytest {posargs}
 
 [testenv:coverage]
-basepython = python3.10
+base_python = py310
+deps = coverage[toml]
+skip_install = true
+parallel_show_output = true
+# Keep in-sync with test env definition above.
+depends = py310-{test}
 commands =
     python -m coverage combine
     python -m coverage report -m --skip-covered
-    python -m coverage json
 
 [gh-actions]
 python =
