+++ .github/workflows/ci.yml
@@ -19,7 +19,6 @@ jobs:
         os:
           - ubuntu-latest
           - macos-latest
-          - windows-latest
         python-version:
           - "3.10"
 
