+++ .github/workflows/ci.yml
@@ -56,7 +56,7 @@ jobs:
       - name: "Set up Python"
         uses: "actions/setup-python@v4"
         with:
-          python-version: "{{ matrix.python-version }}"
+          python-version: "${{ matrix.python-version }}"
 
       - name: "Install dependencies"
         run: |
