+++ .github/workflows/ci.yml
@@ -73,8 +73,8 @@ jobs:
           python -m coverage html --skip-covered --skip-empty
 
           # Report and write to summary.
-          python -m coverage report | sed 's/^/    /' >> $GITHUB_STEP_SUMMARY
-
+          python -Im coverage report --format=markdown >> $GITHUB_STEP_SUMMARY
+          
           # Report again and fail if under 100%.
           python -Im coverage report --fail-under=100
 
