+++ tox.ini
@@ -18,7 +18,7 @@ parallel_show_output = true
 depends = py310-{test}
 commands =
     python -m coverage combine
-    python -m coverage report -m --skip-covered
+    python -m coverage report 
 
 [gh-actions]
 python =
