+++ .github/workflows/ci.yml
@@ -18,7 +18,6 @@ jobs:
       matrix:
         os:
           - ubuntu-latest
-          - macos-latest
         python-version:
           - "3.10"
 
@@ -44,6 +43,7 @@ jobs:
         with:
           name: "coverage-data"
           path: .coverage.*
+          if-no-files-found: ignore
 
   coverage:
     name: "Combine and check coverage"
@@ -74,7 +74,7 @@ jobs:
 
           # Report and write to summary.
           python -Im coverage report --format=markdown >> $GITHUB_STEP_SUMMARY
-          
+
           # Report again and fail if under 100%.
           python -Im coverage report --fail-under=100
 
