+++ .github/workflows/ci.yml
@@ -72,12 +72,6 @@ jobs:
           python -m coverage combine
           python -m coverage html --skip-covered --skip-empty
 
-          # Report and write to summary.
-          python -Im coverage report --format=markdown >> $GITHUB_STEP_SUMMARY
-
-          # Report again and fail if under 100%.
-          python -Im coverage report --fail-under=100
-
       - name: "Upload HTML report if check failed"
         uses: actions/upload-artifact@v3
         with:
