+++ .github/workflows/ci.yml
@@ -60,7 +60,7 @@ jobs:
 
       - name: "Install dependencies"
         run: |
-          python -m pip install --upgrade coverage[toml]
+          python -Im pip install --upgrade coverage[toml]
 
       - name: "Download coverage data"
         uses: actions/download-artifact@v3
@@ -69,8 +69,14 @@ jobs:
 
       - name: "Combine coverage and fail it it’s under 100 %"
         run: |
-          python -m coverage combine
-          python -m coverage html --skip-covered --skip-empty
+          python -Im coverage combine
+          python -Im coverage html --skip-covered --skip-empty
+
+          # Report and write to summary.
+          python -Im coverage report --format=markdown >> $GITHUB_STEP_SUMMARY
+
+          # Report again and fail if under 100%.
+          python -Im coverage report --fail-under=100
 
       - name: "Upload HTML report if check failed"
         uses: actions/upload-artifact@v3
