+++ .github/workflows/ci.yml
@@ -54,7 +54,7 @@ jobs:
       - uses: actions/checkout@v4
       - uses: actions/setup-python@v5
         with:
-          python-version-file: .python-version-default
+          python-version: "${{ matrix.python-version }}"
           cache: pip
       - run: python -Im pip install --upgrade coverage[toml]
       - uses: actions/download-artifact@v4
