+++ .github/workflows/ci.yml
@@ -56,7 +56,7 @@ jobs:
         with:
           python-version: "${{ matrix.python-version }}"
           cache: pip
-      - run: python -Im pip install --upgrade coverage[toml]
+      - run: python -m pip install --upgrade coverage[toml]
       - uses: actions/download-artifact@v4
         with:
           pattern: coverage-data
@@ -64,14 +64,14 @@ jobs:
 
       - name: "Combine coverage and fail it it’s under 100 %"
         run: |
-          python -Im coverage combine
-          python -Im coverage html --skip-covered --skip-empty
+          python -m coverage combine
+          python -m coverage html --skip-covered --skip-empty
 
           # Report and write to summary.
-          python -Im coverage report --format=markdown >> $GITHUB_STEP_SUMMARY
+          python -m coverage report --format=markdown >> $GITHUB_STEP_SUMMARY
 
           # Report again and fail if under 100%.
-          python -Im coverage report --fail-under=100
+          python -m coverage report --fail-under=100
 
       - name: "Upload HTML report if check failed"
         uses: actions/upload-artifact@v3
