+++ .github/workflows/ci.yml
@@ -59,7 +59,7 @@ jobs:
       - run: python -m pip install --upgrade coverage[toml]
       - uses: actions/download-artifact@v4
         with:
-          pattern: coverage-data
+          pattern: coverage-data*
           merge-multiple: true
 
       - name: "Combine coverage and fail it it’s under 100 %"
