+++ .github/workflows/ci.yml
@@ -39,9 +39,9 @@ jobs:
           python -m tox
 
       - name: "Upload coverage data"
-        uses: actions/upload-artifact@v3
+        uses: actions/upload-artifact@v4
         with:
-          name: "coverage-data"
+          name: "coverage-data-${{ matrix.python-version }}"
           path: .coverage.*
           if-no-files-found: ignore
 
@@ -59,7 +59,7 @@ jobs:
       - run: python -m pip install --upgrade coverage[toml]
       - uses: actions/download-artifact@v4
         with:
-          pattern: coverage-data*
+          pattern: coverage-data-*
           merge-multiple: true
 
       - name: "Combine coverage and fail it it’s under 100 %"
