+++ .github/workflows/ci.yml
@@ -70,8 +70,9 @@ jobs:
           # Report and write to summary.
           python -m coverage report --format=markdown >> $GITHUB_STEP_SUMMARY
 
-          # Report again and fail if under 100%.
-          python -m coverage report --fail-under=50
+          # Write total to environment.
+          total=$(python -m coverage report --fail-under=50 --show-missing | grep TOTAL | awk '{print $NF}')
+          echo "total=$total" >> $GITHUB_ENV
 
       - name: "Upload HTML report if check failed"
         uses: actions/upload-artifact@v3
