### 2024-03-26 3180bb7
Adding coverage on ReadMe (#10)
* Create ci.yml

Add coverage badge

* Update ci.yml

* fix tox.ini

* python version in ci.yml

* python version in ci.yml

* python version in ci.yml

* python version in ci.yml

* python version in ci.yml

* python version in ci.yml

* python version in ci.yml

* python version in ci.yml

* python version in ci.yml

* python version in ci.yml

* python version in ci.yml

* python version in ci.yml

* python version in ci.yml

* python version in ci.yml

* python version in ci.yml

* python version in ci.yml

* python version in ci.yml

* python version in ci.
[…]
