+++ tox.ini
@@ -1,0 +1,26 @@
+[tox]
+envlist = py310-{test},coverage
+
+[testenv]
+package = wheel
+wheel_build_env = .pkg
+extras =
+    test: test
+commands =
+    python -m coverage run -p -m pytest {posargs}
+
+[testenv:coverage]
+base_python = py310
+deps = coverage[toml]
+skip_install = true
+parallel_show_output = true
+# Keep in-sync with test env definition above.
+depends = py310-{test}
+commands =
+    python -m coverage combine
+    python -m coverage report 
+
+[gh-actions]
+python =
+    3.10: py310
+
