+++ .github/workflows/ci.yml
@@ -1,0 +1,94 @@
+# Run CI
+
+name: "CI"
+
+on:
+  push:
+    branches: [main, tests]
+  pull_request:
+    branches: [main]
+
+jobs:
+  tests:
+    name: "Python ${{ matrix.python-version }} on ${{ matrix.os }}"
+    runs-on: "${{ matrix.os }}"
+
+    strategy:
+      fail-fast: false
+      matrix:
+        os:
+          - ubuntu-latest
+        python-version:
+          - "3.10"
+
+    steps:
+      - name: "Check out the repo"
+        uses: "actions/checkout@v3"
+
+      - name: "Set up Python"
+        uses: "actions/setup-python@v4"
+        with:
+          python-version: "${{ matrix.python-version }}"
+
+      - name: "Install dependencies"
+        run: |
+          python -m pip install tox tox-gh-actions
+
+      - name: "Run tox for ${{ matrix.python-version }}"
+        run: |
+          python -m tox
+
+      - name: "Upload coverage data"
+        uses: actions/upload-artifact@v4
+        with:
+          name: "coverage-data-${{ matrix.python-version }}"
+          path: .coverage.*
+          if-no-files-found: ignore
+
+  coverage:
+    name: Combine & check coverage
+    needs: tests
+    runs-on: ubuntu-latest
+
+    steps:
+      - uses: actions/checkout@v4
+      - uses: actions/setup-python@v5
+        with:
+          python-version: "${{ matrix.python-version }}"
+          cache: pip
+      - run: python -m pip install --upgrade coverage[toml]
+      - uses: actions/download-artifact@v4
+        with:
+          pattern: coverage-data-*
+          merge-multiple: true
+
+      - name: "Combine coverage and fail it it’s under 100 %"
+        run: |
+          python -m coverage combine
+          python -m coverage html --skip-covered --skip-empty
+
+          # Report and write to summary.
+          python -m coverage report --format=markdown >> $GITHUB_STEP_SUMMARY
+
+          # Write total to environment.
+          total=$(python -m coverage report --fail-under=50 --show-missing | grep TOTAL | awk '{print $NF}')
+          echo "total=$total" >> $GITHUB_ENV
+
+      - name: "Upload HTML report if check failed"
+        uses: actions/upload-artifact@v3
+        with:
+          name: html-report
+          path: htmlcov
+        if: ${{ failure() }}
+
+      - name: "Create badge"
+        uses: schneegans/dynamic-badges-action@v1.6.0
+        with:
+          auth: ${{ secrets.GIST_SECRET }}
+          gistID: 76065cdb0e28e9ce276f32f5c5ac568f
+          filename: covbadge.json
+          label: Coverage
+          message: ${{ env.total }}
+          minColorRange: 50
+          maxColorRange: 90
+          valColorRange: ${{ env.total }}
