### 2024-03-26 #6493 @alixlam
@Kevin-Mattheus-Moerman  thank you for your feedback on the initial steps. I am sorry for the delay in my response. Concerning the license file, I was hoping to have some guidance on that, as they are many licenses possible on the link you provided I am not sure which one to use. On testing, I have implemented some tests for this project, and made sure that the tests are run at each push to the main branch, however maybe there are not enough ? I added today to the repository the information on the test coverage (thanks to a badge on the README), today the coverage is 71%.
