### 2024-03-26 #6493 @Kevin-Mattheus-Moerman
@alixlam Unfortunately we cannot offer advice on a license to choose. So even what I share here is by no means legal advice on a license, I'll just recommend, not as an editor representing JOSS, but as a fellow researcher, some steps to understand these licenses. I would take some time at this point to understand the main types and the consequences/outcomes associated with each license. I know the OSI approved license list is very long, however if you do not have a strong license preference at the moment it is probably best to stick to a widely used license. Of those there are two main types, permissive ones (e.g. MIT, Apache 2.0) and "copy left" ones (e.g. GPL-v3). It is important to understand the difference between the two, and to first decide on that. The biggest impact here is for folks using your work, e.g. the copy left license placing constraints on people using your work in terms of the license they may use. Some see this as a beneficial property of copy-left, as it may ensure that derivatives stay free/open (under the same type of license), while others feel this is negative and that there should be more flexibility (even to make a derivative closed-source/proprietary), which leads them to permissive licenses.
I recommend you read some resources on license types (e.g. https://choosealicense.com/, https://en.wikipedia.org/wiki/Open-source_license, and look for some others yourself) and to start from there.

Once again, this text here is by no means any form of legal advice on licenses, I am afraid the sole responsibility in choosing a license rests with you. Myself and JOSS have no preference in terms of the type of OSI approved license you choose.

Hope this helps.
