### 2024-03-27 #6493 @alixlam
@Kevin-Mattheus-Moerman , thanks a lot for the clarification it definitely helped , I added the LICENSE file to the project.
