### 2024-03-31 #6493 @Kevin-Mattheus-Moerman
@alixlam thanks for adding that license. I have just started looking for an editor to handle this submission. One point that will come up during review is contributing guidelines, which I believe you currently do not have. I recommend you add a `CONTRIBUTING.md` file to your main folder, and to link to it from the README (here are some examples: https://contributing.md/example/).
