### 2024-04-02 #6493 @sappelhoff
Hi @alixlam I will be editing your submission.

Do you have a list of potential reviewers for your project, that I could screen and use for invitations?

Please also remember to respond to this [earlier request](https://github.com/openjournals/joss-reviews/issues/6493#issuecomment-2028605751):

>  Secondly please add the city and country for both affiliations on the paper, thanks.

Finally, why are there a number of authors on the paper that (seemingly) have not committed to the repository?
