### 2024-04-05 #6493 @sappelhoff
>  I do not have a list of potential reviewers unfortunately.

Ok, no problem -- I will start trying to find some. In case you think of a suitable reviewer in the meantime, please let me know. It'd be a great help and speed up the review process.

> I will add this information right away.

Great, thank you.

>  The other authors have indeed not committed to the repository but have guided me in the process, I am a phd student and they are my thesis supervisors.

Thanks for the explanation. For your reference, here are our guidelines as to the "responsibilities" and "criteria" of being an author for a JOSS paper: https://joss.readthedocs.io/en/latest/submitting.html#authorship

Some further points are listed under "ethics" here: https://joss.theoj.org/about#ethics

In case you haven't checked this documentation before, please check it now. If afterwards you decide that the list of authors is fine as it is, then that's fine with me!
