### 2024-04-15 #6493 @mnarayan
I won't be able to for 2 weeks.

On Mon, Apr 15, 2024, 7:41 AM Stefan Appelhoff ***@***.***>
wrote:

> 👋 @a3sha2 <https://github.com/a3sha2> @szorowi1
> <https://github.com/szorowi1> @mnarayan <https://github.com/mnarayan>,
> would any of you be willing to review this submission for JOSS? We carry
> out our checklist-driven reviews here in GitHub issues and follow these
> guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
>
> I will also try to send you a more formal invitation via email.
>
