### 2024-04-16 #6493 @sappelhoff
> Thanks for waiting for my response, I was travelling. I can review it. As a disclaimer: my (f)MRI days are over for some years but I will manage :)

Thanks @behinger, I am planning to assign a team of 3 reviewers, so between all of you, all things will probably be covered!
