### 2024-04-16 #6636 @editorialbot
[REVIEW]: fMRIStroke: A preprocessing pipeline for fMRI Data from Stroke patients
**Submitting author:** @alixlam (<a href="http://orcid.org/0009-0008-2046-3703">Alix Marie Eleonore Lamouroux</a>)
**Repository:** https://github.com/alixlam/fmristroke
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
