### 2024-04-17 alixlam/fmristroke#11 @SRSteinkamp
Installation requirements not working with python 3.8.19
Hi all,

currently, reviewing the toolbox (https://github.com/openjournals/joss-reviews/issues/6636).

During my first attempt at installing fmristroke in Python 3.8.19 (new conda environment), there was an issue with the specific ipython version (8.16.1). These were resolved by using 3.9.

I am not sure what the best way would be to have a requirements.txt file that works best with different python versions, the simplest solution could be to not list python 3.8. Maybe the other are able to provide some more substational input.
