### 2024-04-17 #6636 @SRSteinkamp
As seen above, create an issue regarding the mismatched between the listed installation requirements and python verison: https://github.com/alixlam/fmristroke/issues/11
Maybe someone else has a better idea of good practices writing requirements.txts .
