### 2024-04-17 alixlam/fmristroke#12 @SRSteinkamp
Installation: data/atlas is not installed (thus not available)
It seems to that the data/atlas folder is not probably installed, downloaded when using pip install.

This might be fixed here:
https://github.com/alixlam/fmristroke/blob/22b97a0ddd6f889a761751eb00f83b5505a0e99f/setup.py#L41C1-L41C81

Referencing the review here (https://github.com/openjournals/joss-reviews/issues/6636).
