+++ requirements.txt
@@ -18,8 +18,6 @@ attrs==23.1.0
     # via
     #   niworkflows
     #   sdcflows
-backcall==0.2.0
-    # via ipython
 bids-validator==1.13.1
     # via pybids
 bokeh==2.2.3
@@ -28,7 +26,7 @@ certifi==2023.7.22
     # via requests
 charset-normalizer==3.3.0
     # via requests
-chart-studio==1.1.0
+chart-studio==1.1.
     # via antspyx
 ci-info==0.3.0
     # via etelemetry
@@ -39,24 +37,16 @@ click==8.1.7
     #   pybids
 codecarbon==2.3.1
     # via fmriprep
-comm==0.1.4
-    # via ipykernel
 contourpy==1.1.1
     # via matplotlib
 cycler==0.12.0
     # via matplotlib
 cython==3.0.3
     # via nitime
-debugpy==1.8.0
-    # via ipykernel
-decorator==5.1.1
-    # via ipython
 docopt==0.6.2
     # via num2words
 etelemetry==0.3.0
     # via nipype
-exceptiongroup==1.1.3
-    # via ipython
 executing==2.0.0
     # via stack-data
 filelock==3.12.4
@@ -81,14 +71,8 @@ indexed-gzip==1.8.5
     # via smriprep
 interface-meta==1.3.0
     # via formulaic
-ipykernel==6.20.2
-    # via -r requirements.in
-ipython==8.16.1
-    # via ipykernel
 isodate==0.6.1
     # via rdflib
-jedi==0.19.1
-    # via ipython
 jinja2==3.0.1
     # via
     #   bokeh
@@ -98,8 +82,6 @@ joblib==1.3.2
     # via
     #   nilearn
     #   scikit-learn
-jupyter-client==8.3.1
-    # via ipykernel
 jupyter-core==5.3.2
     # via jupyter-client
 kiwisolver==1.4.5
@@ -132,12 +114,6 @@ matplotlib==3.8.2
     #   seaborn
     #   smriprep
     #   tedana
-matplotlib-inline==0.1.6
-    # via
-    #   ipykernel
-    #   ipython
-nest-asyncio==1.5.8
-    # via ipykernel
 networkx==3.1
     # via
     #   nipype
@@ -222,7 +198,6 @@ packaging==23.2
     # via
     #   bokeh
     #   fmriprep
-    #   ipykernel
     #   matplotlib
     #   nibabel
     #   nilearn
@@ -249,10 +224,6 @@ parso==0.8.3
     # via jedi
 patsy==0.5.3
     # via statsmodels
-pexpect==4.8.0
-    # via ipython
-pickleshare==0.7.5
-    # via ipython
 pillow==10.0.1
     # via
     #   antspyx
@@ -266,15 +237,12 @@ plotly==5.17.0
     # via chart-studio
 prometheus-client==0.17.1
     # via codecarbon
-prompt-toolkit==3.0.39
-    # via ipython
 prov==2.0.0
     # via nipype
 psutil==5.9.5
     # via
     #   codecarbon
     #   fmriprep
-    #   ipykernel
 ptyprocess==0.7.0
     # via pexpect
 pure-eval==0.2.2
@@ -292,8 +260,6 @@ pydot==1.4.2
     # via nipype
 pyfftw==0.13.1
     # via rapidtide
-pygments==2.16.1
-    # via ipython
 pynvml==11.5.0
     # via codecarbon
 pyparsing==3.1.1
@@ -328,7 +294,6 @@ pyyaml==6.0.1
     #   smriprep
 pyzmq==25.1.1
     # via
-    #   ipykernel
     #   jupyter-client
 rapidtide==2.6.5
     # via -r requirements.in
[…]
