+++ requirements.in
@@ -1,4 +1,3 @@
-ipykernel==6.20.2
 antspyx==0.3.8
 fmriprep==23.0.2
 rapidtide
