### 2024-04-23 beabee3
issues #12 #11 installation (#13)
Co-authored-by: Alix LAMOUROUX <a19lamou@sl-mee-br-115.imta.fr>
