+++ requirements.txt
@@ -26,7 +26,7 @@ certifi==2023.7.22
     # via requests
 charset-normalizer==3.3.0
     # via requests
-chart-studio==1.1.
+chart-studio==1.1.0
     # via antspyx
 ci-info==0.3.0
     # via etelemetry
