### 2024-04-23 alixlam/fmristroke#12 @alixlam
Thank you very much @SRSteinkamp , it was indeed the source of the problem thank you for pointing it out. Should be installed now when installing the toolbox.
