### 2024-04-29 alixlam/fmristroke#14 @behinger
Minor issues
Hi! Nice toolbox! I'm a reviewer for JOSS (openjournals/joss-reviews#6636)

Someminor issues I noticed
- [x] "usage notes": "_installation" link doesnt work
![grafik](https://github.com/alixlam/fmristroke/assets/10183650/c9023ae2-3b53-429c-aa50-f675772f8314)
- [x] "changes" is empty
- [x] Python 3.8 still recommended (https://fmristroke.readthedocs.io/en/latest/installation.html) but #11 recommends 3.9
- [x] with Python 3.12 I run into an issue with bokeh
![grafik](https://github.com/alixlam/fmristroke/assets/10183650/ebeaeede-68e3-4c98-8a64-0ffc263615d5)
- [x]  the tests / CI runs on python 3.10 - maybe it is worth to just recommend that version?
- [x]  the scripts folder is empty

all those points should be trivial to adress.

- [x] I was now trying to find a suitable dataset to test out the pipeline, but I'm not at all experienced with stroke, and I dont necessarily want to run fmriprep on an unknown dataset and deal with all those issues. Could you recommend an already processed dataset?

If there is no such dataset readily available, a pointer towards a suitable one to run via fmriprep would be appreciated.

Is it correct to say that your tests do not test the full capacities of the toolbox? At least I havent found any paths/artefacts or so to a dataset. Pointers are appreciated :)
