### 2024-04-29 alixlam/fmristroke#14 @SRSteinkamp
Regarding the last point - I tested running fmriprep with a fake-lesion mask on some data and getting errors -, I think both providing a test dataset, and with that an example of how to prepare your BIDS data (where to put the lesion mask etc.) would be really helpful.
