### 2024-04-30 alixlam/fmristroke#14 @alixlam
Hi @behinger and @SRSteinkamp , thank you for your valuable feedbacks. Sharing the data that I currently use is a bit tricky but I found some open source data and am currently working on adding those as a test dataset. There is actually data used for the tests in "data/tests" but it is completely fake data just to make sure that the code runs. You are also right the documentation lacks some info on the BIDS data preparation, will add this shortly! Thanks again for your time.
