+++ docs/changes.rst
@@ -1,6 +1,0 @@
-
-.. _changes:
-
---------
-Changes
---------
