+++ docs/installation.rst
@@ -1,15 +1,15 @@
 
-.. _installation:
+.. _Installation:
 
 -------------
 Installation
 -------------
 
 
-Manually Prepared Environment (Python 3.8+)
+Manually Prepared Environment (Python 3.9+)
 ===========================================
 
-On a functional Python 3.8 (or above) environment with ``pip`` installed,
+On a functional Python 3.9 (or above) environment with ``pip`` installed,
 *fMRIStroke* can be installed using the habitual command ::
 
     $ python -m pip install git+https://github.com/alixlam/fmristroke.git 
@@ -18,7 +18,7 @@ On a functional Python 3.8 (or above) environment with ``pip`` installed,
 
 External Dependencies
 ---------------------
-*fMRIStroke* is written using Python 3.8 (or above), and is based on
+*fMRIStroke* is written using Python 3.9 (or above), and is based on
 nipype_.
 
 Containerized execution (Docker and Singularity)
