+++ .github/workflows/main.yml
@@ -5,7 +5,7 @@ name: Pytest
 
 on:
   push:
-    branches: [ 'develop', 'main' ]
+    branches: [ 'minor_issues', 'main' ]
   pull_request:
     branches: [ 'develop', 'main' ]
 
@@ -28,7 +28,10 @@ jobs:
     - name: Set up Python 3.10
       uses: actions/setup-python@v3
       with:
-        python-version: '3.10'
+        python-version:
+          -3.10
+          -3.9
+          -3.11
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
