+++ .github/workflows/main.yml
@@ -29,9 +29,9 @@ jobs:
       uses: actions/setup-python@v3
       with:
         python-version:
-          -3.10
-          -3.9
-          -3.11
+          - 3.10
+          - 3.9
+          - 3.11
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
