+++ .github/workflows/main.yml
@@ -22,16 +22,14 @@ jobs:
     strategy:
       matrix:
         os: [ubuntu-latest]
+        python-version: [3.9, 3.10, 3.11, 3.12]
 
     steps:
     - uses: actions/checkout@v3
     - name: Set up Python 3.10
       uses: actions/setup-python@v3
       with:
-        python-version:
-          - 3.10
-          - 3.9
-          - 3.11
+        python-version: ${{ matrix.python-version }}
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
