+++ docs/conf.py
@@ -159,10 +159,6 @@ html_theme = "sphinx_rtd_theme"
 # Add any paths that contain custom themes here, relative to this directory.
 # html_theme_path = []
 
-# The name for this set of Sphinx documents.
-# "<project> v<release> documentation" by default.
-# html_title = u'fmriprep vversion'
-
 # A shorter title for the navigation bar.  Default is the same as html_title.
 # html_short_title = None
 
