+++ .github/workflows/main.yml
@@ -5,9 +5,9 @@ name: Pytest
 
 on:
   push:
-    branches: [ 'develop', 'main' ]
+    branches: [ 'dev', 'main' ]
   pull_request:
-    branches: [ 'develop', 'main' ]
+    branches: [ 'dev', 'main' ]
 
 permissions:
   contents: read
@@ -22,13 +22,14 @@ jobs:
     strategy:
       matrix:
         os: [ubuntu-latest]
+        python-version: ['3.9', '3.10', '3.11']
 
     steps:
     - uses: actions/checkout@v3
     - name: Set up Python 3.10
       uses: actions/setup-python@v3
       with:
-        python-version: '3.10'
+        python-version: ${{ matrix.python-version }}
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
