### 2024-05-01 alixlam/fmristroke#14 @SRSteinkamp
Thanks :)

A few further very minor issues - in the documentation (usage notes):

> fmristroke data/bids_root/ out/ participant -w work/ --fmriprep-dir derivatives/

The argument should be `--fmriprep_dir`.

There's also a typo on the first page here: https://fmristroke.readthedocs.io/en/latest/index.html, saying "fmristoke".

Lastly, in the README, I would move the link to readthedocs a little more up, so that it's easier accessible (and maybe add it at the GitHub repos website).
