### 2024-05-08 alixlam/fmristroke#16 @SRSteinkamp
Minor suggested edits to the paper
Hi again,

reviewing for JOSS (https://github.com/openjournals/joss-reviews/issues/6636) and I have a few very minor suggestions regarding the paper:

The last lines of the abstract (14 - 19), I feel, are a bit incomplete / confusing. I think very minor edits, would make them a lot more powerful.

I'd further suggest to use "quality control" instead of "quality checks", throughout the paper - this might be only me, but "quality checks plots" does not feel right for me.

Citation in line 41 is not correctly formatted.

In line 57, I think there is a "data" missing (as in "fMRI data").
