### 2024-05-13 #6636 @behinger
Hi! Thanks for asking. I think we are both waiting for alixlam to post some test-data / procedure. https://github.com/alixlam/fmristroke/issues/14#issuecomment-2084615050 to get us up to speed.
