### 2024-06-02 #6636 @sappelhoff
Hey @alixlam 👋 could you please supply a short update on your efforts on adding some test-data / procedure?
