### 2024-06-12 #6636 @alixlam
Hello, so sorry for the time taken to respond. I did find open data to share and ran fMRIprep on it to have the necessary derivatives to test the repository. However, I am not sure it is a good idea to add the data files directly to the repository as they are quite heavy files. Do you have any suggestions on how to share them ?  @sappelhoff @behinger
