### 2024-06-12 #6636 @behinger
You could upload them to osf (or fig share if you prefer), and link it. Then write a short documentation page how to access+run these data. Typically open euro is cc-by, so you can easily rehost.

I would not necessarily run a system unit-test with them if they cannot be made smaller.
