### 2024-07-17 alixlam/fmristroke#17 @SRSteinkamp
Potential bug: not all expectations of fMRIPrep  were met
Hi,

I hope you can help me trouble shoot the following issue. In both using the data @alixlam provided on OSF and in using some data of my own with a fake lesion mask I am getting this error.

It works, however, using the test data, that is somewhere stored in the installation. An as far as I can see there are no obvious differences between the OSF data and the example data. Except that the osf data also contains T2w images.

```
Attempted to access pre-existing anatomical derivatives at         <PATH/osfstorage-archive/ARC - derivatives>, however not all expectations of fMRIPrep         were met (for participant <M2112>, spaces <MNI152NLin2009cAsym>,         >).
Process Process-2:
Traceback (most recent call last):
  File "~/anaconda3/envs/joss/lib/python3.9/multiprocessing/process.py", line 315, in _bootstrap
    self.run()
  File "/~/anaconda3/envs/joss/lib/python3.9/multiprocessing/process.py", line 108, in run
    self._target(*self._args, **self._kwargs)
  File "~/anaconda3/envs/joss/lib/python3.9/site-packages/fmristroke/cli/workflow.py", line 73, in build_workflow
    retval["workflow"] = init_fmristroke_wf()
  File "~/anaconda3/envs/joss/lib/python3.9/site-packages/fmristroke/workflows/final.py", line 42, in init_fmristroke_wf
    single_subject_wf = init_single_subject_wf(subject_id)
  File "~/anaconda3/envs/joss/lib/python3.9/site-packages/fmristroke/workflows/final.py", line 146, in init_single_subject_wf
    BIDSDerivativeDataGrabber(
  File "/~/anaconda3/envs/joss/lib/python3.9/site-packages/nipype/interfaces/base/core.py", line 563, in __init__
    super(SimpleInterface, self).__init__(
  File "~/anaconda3/envs/joss/lib/python3.9/site-packages/nipype/interfaces/base/core.py", line 202, in __init__
    self.inputs.trait_set(**inputs)
  File "~/anaconda3/envs/joss/lib/python3.9/site-packages/traits/has_traits.py", line 1520, in trait_set
    setattr(self, name, value)
  File "~/anaconda3/envs/joss/lib/python3.9/site-packages/traits/trait_types.py", line 3101, in validate
    self.error(object, name, value)
  File "~/anaconda3/envs/joss/lib/python3.9/site-packages/traits/base_trait_handler.py", line 74, in error
    raise TraitError(
traits.trait_errors.TraitError: The 'anat_derivatives' trait of a _BIDSDerivativeDataGrabberInputSpec instance must be a dictionary with keys which are a string and with values which are any value, but a value of None <class 'NoneType'> was specified.
```

Do you have any idea where this issues might be coming from? In any case, a slightly more verbose error message mit be helpful at this stage, cause I am not sure what exactly is missing from fmriprep.

Cheers

PS:
reviewing for JOSS (https://github.com/openjournals/joss-reviews/issues/6636)
