### 2024-07-18 alixlam/fmristroke#17 @alixlam
hi @SRSteinkamp ! Thank you for your message ! I just found the error the sub-M2112 folder was missing 1 file ... I just added it  ("sub-M2112_space-MNI152NLin2009cAsym_label-WM_probseg.nii.gz"). On your data I am not sure, have you run fmriprep with the freesurfer option ? I will work on something to add verbose in the error msg .
