### 2024-07-30 cdf173e
Update paper.md
