### 2024-08-08 alixlam/fmristroke#17 @SRSteinkamp
Thank you!

Regarding my own data, can we maybe check, what fmriprep version you would recommend and which options? Afaik I ran using freesurfer, but that also has a few options ;)

So I ran the code using default settings on the example data and test data. Both terminate with errors.

On the example data the toolbox now terminates with an error related to rapitide (I guess)

```python
240807-18:35:07,777 nipype.workflow CRITICAL:
	 fMRIStroke failed: Traceback (most recent call last):
  File "~/anaconda3/envs/joss/lib/python3.9/site-packages/nipype/pipeline/plugins/multiproc.py", line 67, in run_node
    result["result"] = node.run(updatehash=updatehash)
  File "~/anaconda3/envs/joss/lib/python3.9/site-packages/nipype/pipeline/engine/nodes.py", line 527, in run
    result = self._run_interface(execute=True)
  File "~/anaconda3/envs/joss/lib/python3.9/site-packages/nipype/pipeline/engine/nodes.py", line 645, in _run_interface
    return self._run_command(execute)
  File "~/anaconda3/envs/joss/lib/python3.9/site-packages/nipype/pipeline/engine/nodes.py", line 771, in _run_command
    raise NodeExecutionError(msg)
nipype.pipeline.engine.nodes.NodeExecutionError: Exception raised while executing Node rapidtide.

Cmdline:
	rapidtide PATH/joss_review/work/fmristroke_0_1_wf/single_subject_M2121_wf/lesion_conn_ses_258_task_rest_run_17_space_T1w_desc_preproc_wf/concat_runs_wf/concat_bold_denoised_t1/concat_bold_pipe_3_space_0.nii.gz PATH/joss_review/work/fmristroke_0_1_wf/single_subject_M2121_wf/lesion_conn_ses_258_task_rest_run_17_space_T1w_desc_preproc_wf/hemodynamic_wf/rapidtide/concat --corrmask PATH/joss_review/work/fmristroke_0_1_wf/single_subject_M2121_wf/lesion_conn_ses_258_task_rest_run_17_space_T1w_desc_preproc_wf/hemodynamic_wf/intersect_mask/mask_inter.nii.gz --filterband None --globalmeanexclude PATH/joss_review/work/fmristroke_0_1_wf/single_subject_M2121_wf/lesion_conn_ses_258_task_rest_run_17_space_T1w_desc_preproc_wf/hemodynamic_wf/roi_resamp/mask_resamp.nii.gz --globalmeaninclude PATH/joss_review/work/fmristroke_0_1_wf/single_subject_M2121_wf/lesion_conn_ses_258_task_rest_run_17_space_T1w_desc_preproc_wf/hemodynamic_wf/intersect_mask/mask_inter.nii.gz --motpos --noglm --searchrange -10 10 --spatialfilt 3.000000
Stdout:
	memprofiler does not exist
	no aggressive optimization
	will not use numba even if present
	startpoint set to minimum, (0)
	endppoint set to maximum, ( 9 )
Stderr:
	starting rapidtide 2.6.5
	running single process - disabled shared memory use
	setting internal precision to double
	setting output precision to single
	input file is NIFTI
	oversample factor set to 2
	fmri data: 10 timepoints, tr = 1.0, oversamptr = 0.5
	276696 spatial locations, 10 timepoints
	Traceback (most recent call last):
	  File "~/anaconda3/envs/joss/bin/rapidtide", line 23, in <module>
	    rapidtide_workflow.rapidtide_main(rapidtide_parser.process_args(inputargs=None))
	  File "~/anaconda3/envs/joss/lib/python3.9/site-packages/rapidtide/workflows/rapidtide.py", line 425, in rapidtide_main
	    raise ValueError(
	ValueError: magnitude of lagmin exceeds 5.0 - invalid
Traceback:
	Traceback (most recent call last):
	  File "~/anaconda3/envs/joss/lib/python3.9/site-packages/nipype/interfaces/base/core.py", line 453, in aggregate_outputs
	    setattr(outputs, key, val)
	  File "~/anaconda3/envs/joss/lib/python3.9/site-packages/nipype/interfaces/base/traits_extension.py", line 330, in validate
	    value = super(File, self).validate(objekt, name, value, return_pathlike=True)
	  File "~/anaconda3/envs/joss/lib/python3.9/site-packages/nipype/interfaces/base/traits_extension.py", line 135, in validate
	    self.error(objekt, name, str(value))
	  File "~/anaconda3/envs/joss/lib/python3.9/site-packages/traits/base_trait_handler.py", line 74, in error
	    raise TraitError(
	traits.trait_errors.TraitError: The 'output_corrfit_mask' trait of a RapidTideOutputSpec instance must be a pathlike object or string representing an existing file, but a value of 'PATH/joss_review/work/fmristroke_0_1_wf/single_subject_M2121_wf/lesion_conn_ses_258_task_rest_run_17_space_T1w_desc_preproc_wf/hemodynamic_wf/rapidtide/concat_desc-corrfit_mask.nii.gz' <class 'str'> was specified.

	Traceback (most recent call last):
	  File "~/anaconda3/envs/joss/lib/python3.9/site-packages/nipype/interfaces/base/core.py", line 400, in run
	    outputs = self.aggregate_outputs(runtime)
	  File "~/anaconda3/envs/joss/lib/python3.9/site-packages/nipype/interfaces/base/core.py", line 460, in aggregate_outputs
	    raise FileNotFoundError(msg)
	FileNotFoundError: No such file or directory 'PATH/joss_review/work/fmristroke_0_1_wf/single_subject_M2121_wf/lesion_conn_ses_258_task_rest_run_17_space_T1w_desc_preproc_wf/hemodynamic_wf/rapidtide/concat_desc-corrfit_mask.nii.gz' for output 'output_corrfit_mask' of a RapidTide interface
```

And running on the data fails due to a bunch of error I think related to ICA and empty masks:

```python
nipype.pipeline.engine.nodes.NodeExecutionError: Exception raised while executing Node acc_masks.

Traceback:
	Traceback (most recent call last):
	  File "~/anaconda3/envs/joss/lib/python3.9/site-packages/nipype/interfaces/base/core.py", line 397, in run
	    runtime = self._run_interface(runtime)
	  File "~/anaconda3/envs/joss/lib/python3.9/site-packages/fmriprep/interfaces/confounds.py", line 83, in _run_interface
	    self._results["out_masks"] = acompcor_masks(
	  File "~/anaconda3/envs/joss/lib/python3.9/site-packages/fmriprep/utils/confounds.py", line 133, in acompcor_masks
	    csf_file = mask2vf(
	  File "~/anaconda3/envs/joss/lib/python3.9/site-packages/fmriprep/utils/confounds.py", line 48, in mask2vf
	    max_data = np.percentile(data[data > 0], 99)
	  File "~/anaconda3/envs/joss/lib/python3.9/site-packages/numpy/lib/function_base.py", line 4283, in percentile
	    return _quantile_unchecked(
	  File "~/anaconda3/envs/joss/lib/python3.9/site-packages/numpy/lib/function_base.py", line 4555, in _quantile_unchecked
	    return _ureduce(a,
	  File "~/anaconda3/envs/joss/lib/python3.9/site-packages/numpy/lib/function_base.py", line 3823, in _ureduce
	    r = func(a, **kwargs)
	  File "~/anaconda3/envs/joss/lib/python3.9/site-packages/numpy/lib/function_base.py", line 4721, in _quantile_ureduce_func
	    result = _quantile(arr,
	  File "~/anaconda3/envs/joss/lib/python3.9/site-packages/numpy/lib/function_base.py", line 4830, in _quantile
	    slices_having_nans = np.isnan(arr[-1, ...])
	IndexError: index -1 is out of bounds for axis 0 with size 0
```
