### 2024-08-21 alixlam/fmristroke#17 @alixlam
hello again @SRSteinkamp ! Sorry for the late reply.

Concerning the fmriprep version used, I use the one that is in the requirements file (23.0.x). I believe the derivatives should be the same for the other versions though ... but will figure something out to be able to know exactly which file is missing. For FreeSurfer I used the default in fmriprep without changing anything.

For the first error, it is another "problem" in the data I provided ... The run 17 of patient 2121 is actually too short (only 10 volumes) for the rapid tide part. When I tested the data, I thought I used the default parameters but I only tested with --session-level option which concatenates the runs of the same session and task making a new run that is long enough.
I will add something to warn users about this limit in size of functional runs.

For the second error, you probably are right. On which data are you trying to run the package ?
