### 2024-08-22 alixlam/fmristroke#17 @SRSteinkamp
I think the second error was on the test data (test data from within the python installation).

Good to know, will try to give it a new run on my own data using fmriprep 23.0.x, and using --session-level for the OSF data.
