### 2024-08-23 #6636 @sappelhoff
Hey @alixlam I can see some recent activity in https://github.com/alixlam/fmristroke/issues/17#issuecomment-2301513168

Could you please report on your expected timeline to address other issues raised by the reviewers, such as:

- https://github.com/alixlam/fmristroke/issues/16
- https://github.com/alixlam/fmristroke/issues/14
