+++ fmristroke/utils/bids.py
@@ -5,7 +5,76 @@ from pathlib import Path
 
 from bids import BIDSLayout
 from bids.layout import Query
+from bids.layout.writing import build_path
 from pkg_resources import resource_filename as pkgrf
+
+def collect_anat_derivatives(
+    derivatives_dir, subject_id, std_spaces, freesurfer, spec=None, patterns=None
+):
+    """Gather existing derivatives and compose a cache."""
+    if spec is None or patterns is None:
+        _spec, _patterns = tuple(
+            loads(Path(pkgrf("fmristroke", "data/io_spec_anat.json")).read_text()).values()
+        )
+
+        if spec is None:
+            spec = _spec
+        if patterns is None:
+            patterns = _patterns
+
+    derivs_cache = defaultdict(list, {})
+    derivatives_dir = Path(derivatives_dir)
+
+    def _check_item(item):
+        if not item:
+            return None
+
+        if isinstance(item, str):
+            item = [item]
+
+        result = []
+        for i in item:
+            if not (derivatives_dir / i).exists():
+                i = i.rstrip(".gz")
+                if not (derivatives_dir / i).exists():
+                    return result
+            result.append(str(derivatives_dir / i))
+
+        return result
+
+    for space in [None] + std_spaces:
+        for k, q in spec["baseline"].items():
+            q["subject"] = subject_id
+            if space is not None:
+                q["space"] = space
+            item = _check_item(build_path(q, patterns, strict=True))
+            if space:
+                derivs_cache["std_%s" % k] += item if len(item) == 1 else [item]
+            else:
+                derivs_cache["t1w_%s" % k] = item[0] if len(item) == 1 else item
+
+    for space in std_spaces:
+        for k, q in spec["std_xfms"].items():
+            q["subject"] = subject_id
+            q["from"] = q["from"] or space
+            q["to"] = q["to"] or space
+            item = _check_item(build_path(q, patterns))
+            if not item:
+                return None
+            derivs_cache[k] += item
+
+    derivs_cache = dict(derivs_cache)  # Back to a standard dictionary
+
+    if freesurfer:
+        for k, q in spec["surfaces"].items():
+            q["subject"] = subject_id
+            item = _check_item(build_path(q, patterns))
+            if len(item) == 1:
+                item = item[0]
+            derivs_cache[k] = item
+
+    derivs_cache["template"] = std_spaces
+    return derivs_cache
 
 
 def collect_bold_derivatives(
