### 2024-08-26 82865da
Add warnning when size of bold not long enough
