### 2024-08-26 20564ca
Add warnning when size of bold not long enough
