+++ paper/paper.md
@@ -35,8 +35,7 @@ bibliography: paper.bib
 
 # Summary
 Functional Magnetic Resonance Imaging (fMRI) is a widely used neuroimaging technique for the analysis of neural activity and functional connectivity. However, fMRI signal is inherently noisy and susceptible to various artifacts, compromising the accuracy and reliability of derived analyses. This becomes particularly critical when dealing with stroke patients, given the added complexitiy associated with their neurological condition. Specific preprocessing and denoising are integral steps to identify the nuisance sources and mitigate their effect on fMRI analysis. 
-To address these considerations, we introduce fMRIStroke, a functional magnetic resonance imaging (fMRI) data quality controls and preprocessing pipeline tailored specifically for stroke data. fMRIStroke works as an add-on to typical preprocessing pipeline. Leveraging on a commonly used fMRI preprocessing tool outputs (fMRIprep), fMRIStroke generates new quality control plots, computes additional confounds, and executes confound regression (denoising), providing preprocessed fMRI ready for neural activity or connectivity analysis. 
-
+To address these challenges, we present fMRIStroke, a dedicated preprocessing pipeline designed specifically for the quality control and preprocessing of fMRI data from stroke patients. fMRIStroke operates as an enhancement to standard preprocessing workflows. Building on the outputs from commonly used tools like fMRIprep, fMRIStroke introduces additional quality control visualizations, computes supplementary confounding variables, and performs confound regression (denoising), resulting in preprocessed fMRI data that is ready for subsequent analysis of neural activity or connectivity.
 
 
 # Statement of need
@@ -50,13 +49,13 @@ Existing tools like fMRIprep [@Esteban:2019] have streamlined most of the prepro
 To this end, we propose fMRIStroke, a fMRI preprocessing pipeline designed specifically for stroke data that builds on the output derivatives of fMRIprep [@Esteban:2019] when a lesion mask is given. It provides an interface requiring minimal user input while delivering easily interpretable and comprehensive reports. 
 
 
-Concretely, novel quality controls include a hemodynamic lagmap. fMRIStroke assesses hemodynamic lag using cross-correlation with the global gray matter signal [@Siegel:2016] using the rapidTide  python library [@Rapidtide]. As stroke can introduce altered blood flow patterns a normal hemodynamic response cannot be assumed. [@Siegel:2016] investigated temporal delays (lag) in resting state functional magnetic resonance imaging signals and found that significant hemodynamic lag was observed in 30% of sub-acute stroke patients and approximately 10% of patients at one-year post-stroke. 
+Concretely, novel quality controls include a hemodynamic lagmap. fMRIStroke assesses hemodynamic lag using cross-correlation with the global gray matter signal [@Siegel:2016] using the rapidTide  python library [@Rapidtide]. As stroke can introduce altered blood flow patterns a normal hemodynamic response cannot be assumed. @Siegel:2016 investigated temporal delays (lag) in resting state functional magnetic resonance imaging signals and found that significant hemodynamic lag was observed in 30% of sub-acute stroke patients and approximately 10% of patients at one-year post-stroke. 
 Lags systematically alter measurements of functional connectivity from the affected nodes, and need to be taken into consideration when doing FC analysis [@Siegel:2017].
 
[…]
