### 2024-08-27 28e0828
Add verbose when not all necessary derivatives present #17  (#18)
* Add verbose on finding anat derivatives

* Add warnning when size of bold not long enough

* Add warnning when size of bold not long enough

* sample data change

* sample data change

* sample data change

---------

Co-authored-by: Alix LAMOUROUX <a19lamou@sl-mee-br-115.imta.fr>
