### 2024-08-27 alixlam/fmristroke#16 @alixlam
Thank you for your suggested edits. I changed the part in the abstract and hope it is a bit better. And corrected the little errors here and there. :)
