### 2024-09-02 alixlam/fmristroke#17 @SRSteinkamp
Hey, thanks for all the add-ons.

I finally got it to run on my own data :) ... The issue seems to be, that fmristroke is not looking for anat in session folders.

I.e. I had my data ordered as:

```
sub-001
    ses-01
        anat
        func
```

and fmristroke seems to build fixed paths assuming no session identifier in the name or path.

Not sure if you want to change the data loading completely, but there should be at least a note regarding data organization that there should only be an anat folder in the top directory.

The session parameter might be more important though, once people want to use it for longitudinal data.
