### 2024-09-02 1e81c44
tests generate proper boilerplate
