+++ fmristroke/data/reports-specs.yml
@@ -225,3 +225,28 @@ sections:
 - name: About
   reportlets:
   - bids: {datatype: figures, desc: about, suffix: T1w}
+- name: About
+  nested: true
+  reportlets:
+  - bids: {datatype: figures, desc: about, suffix: T1w}
+  - custom: boilerplate
+    path: '{out_dir}/logs'
+    bibfile: ['fmriprep', 'data/boilerplate.bib']
+    caption: |
+      <p>We kindly ask to report results preprocessed with this tool using the following boilerplate.</p>
+      <p class="alert alert-info" role="alert">
+      <strong>Copyright Waiver</strong>.
+      The boilerplate text was automatically generated by <em>NiReports</em> with the
+      express intention that users should copy and paste this text into their manuscripts <em>unchanged</em>.
+      It is released under the
+      <a href="https://creativecommons.org/publicdomain/zero/1.0/" target="_blank">CC0 license</a>.
+      </p>
+    title: Methods fMRIprep
+  - custom: boilerplate
+    path: '{out_dir}/logsfmristroke'
+    bibfile: ['fmristroke', 'data/boilerplate.bib']
+    title: Methods fMRIstroke
+  - custom: errors
+    path: '{out_dir}/sub-{subject}/log/{run_uuid}'
+    captions: <em>NiReports</em> may have recorded failure conditions.
+    title: Errors
