+++ .github/workflows/ci.yml
@@ -44,6 +44,7 @@ jobs:
           name: "coverage-data-${{ matrix.python-version }}"
           path: .coverage.*
           if-no-files-found: ignore
+          include-hidden-files: true
 
   coverage:
     name: Combine & check coverage
