### 2024-10-08 ba436a8
#17 change fetch derivatives (#19)
* change fetch derivatives

* Update draft-pdf.yml

* Update tox.ini

* Update ci.yml

---------

Co-authored-by: Alix LAMOUROUX <a19lamou@sl-mee-br-115.imta.fr>
