### 2024-10-23 alixlam/fmristroke#20 @SRSteinkamp
Missing community guidelines
Hi,

regarding the JOSS reviewer checklist (https://github.com/openjournals/joss-reviews/issues/6636), the repository is still missing community / contribution guidelines:

Cheers
Simon
