### 2024-10-23 47e02c1
Create CONTRIBUTING.md
add Contributing guidelines #20
