### 2024-11-04 #6636 @sappelhoff
> hey! Sorry I missed that update two weeks ago - I still need to actual check the code with the now provided dataset. Very sorry @alixlam for the delay that this will cause.

@behinger could you give us a tentative timeline on when you aim to complete this?
