### 2024-11-08 #6636 @SRSteinkamp
I ran it using the example data — which works. I had troubles though getting my data into format, as it was only looking for `sub-XX/anat/anat_T1w.nii.gz`, so not being flexible with sessions.
I noticed, though, that I tried using it on a very narrow FOV dataset, which might not go well using ICA etc. (tons of errors there).

Going to try to run it on a new dataset of mine now.
