### 2024-11-09 33f6291
Fix error message when skipping because too short
