### 2024-11-09 e0eacf9
Fix error message when skipping because too short (#22)
Co-authored-by: Alix LAMOUROUX <a19lamou@sl-mee-br-115.imta.fr>
