### 2024-11-12 59893e8
Add Boilerplate for fmristroke (#23)
