+++ fmristroke/workflows/bold/denoise.py
@@ -103,7 +103,7 @@ def init_denoise_wf(
     workflow = Workflow(name=name)
 
     workflow.__desc__ = f"""
-    The BOLD time-series were denoised using the following strategies \n:
+    The BOLD time-series were denoised using the following strategies :\n
     """
     for pipeline in pipelines.pipelines:
         workflow.__desc__ += f"""*{pipeline.pipeline}: {pipeline._desc} \n"""
