+++ fmristroke/cli/workflow.py
@@ -92,3 +92,74 @@ def build_workflow(config_file, retval):
     )
     retval["return_code"] = 0
     return retval
+
+def build_boilerplate(config_file, workflow):
+    """Write boilerplate in an isolated process."""
+    from fmristroke import config
+
+    config.load(config_file)
+    logs_path = config.execution.output_dir / "logs"
+    boilerplate = workflow.visit_desc()
+    citation_files = {
+        ext: logs_path / ("CITATION.%s" % ext) for ext in ("bib", "tex", "md", "html")
+    }
+
+    if boilerplate:
+        # To please git-annex users and also to guarantee consistency
+        # among different renderings of the same file, first remove any
+        # existing one
+        for citation_file in citation_files.values():
+            try:
+                citation_file.unlink()
+            except FileNotFoundError:
+                pass
+
+    citation_files["md"].write_text(boilerplate)
+
+    if not citation_files["md"].exists():
+        from pathlib import Path
+        from subprocess import CalledProcessError, TimeoutExpired, check_call
+
+        from pkg_resources import resource_filename as pkgrf
+
+        bib_text = Path(pkgrf("fmristroke", "data/boilerplate.bib")).read_text()
+        citation_files["bib"].write_text(
+            bib_text.replace("fMRIStroke <version>", f"fMRIStroke {config.environment.version}")
+        )
+
+        # Generate HTML file resolving citations
+        cmd = [
+            "pandoc",
+            "-s",
+            "--bibliography",
+            str(citation_files["bib"]),
+            "--citeproc",
+            "--metadata",
+            'pagetitle="fMRIPrep citation boilerplate"',
+            str(citation_files["md"]),
+            "-o",
+            str(citation_files["html"]),
+        ]
+
+        config.loggers.cli.info("Generating an HTML version of the citation boilerplate...")
+        try:
+            check_call(cmd, timeout=10)
+        except (FileNotFoundError, CalledProcessError, TimeoutExpired):
+            config.loggers.cli.warning("Could not generate CITATION.html file:\n%s", " ".join(cmd))
+
+        # Generate LaTex file resolving citations
+        cmd = [
+            "pandoc",
+            "-s",
+            "--bibliography",
+            str(citation_files["bib"]),
+            "--natbib",
+            str(citation_files["md"]),
+            "-o",
+            str(citation_files["tex"]),
+        ]
+        config.loggers.cli.info("Generating a LaTeX version of the citation boilerplate...")
+        try:
+            check_call(cmd, timeout=10)
+        except (FileNotFoundError, CalledProcessError, TimeoutExpired):
+            config.loggers.cli.warning("Could not generate CITATION.tex file:\n%s", " ".join(cmd))
