+++ fmristroke/data/reports-specs.yml
@@ -223,5 +223,6 @@ sections:
       in green; noise components in red.
     subtitle: ICA Components classified by AROMA
 - name: About
+  nested: true
   reportlets:
   - bids: {datatype: figures, desc: about, suffix: T1w}

[…]
