### 2024-11-12 d467025
Update usage.rst #21
Provide info on fmriprep version compatible.
