### 2024-11-15 #6636 @behinger
I'm happy - good job @alixlam!

Caveat: Because I'm not familiar with stroke-analyses, I cannot confirm the functionality of the analyses in great detail.
