### 2024-11-16 #6636 @sappelhoff
@alixlam please let me know once you finished your items from the list above: https://github.com/openjournals/joss-reviews/issues/6636#issuecomment-2480766557

I will then work on my final items, and then it's (hopefully) finally time to ship it!
