### 2024-11-18 #6636 @sappelhoff
Thanks @alixlam. Here are some more requests from my side:

- [ ] In your archive, there is a directory called `__MACOSX`, please remove it if possible
- [ ] Please can you provide the ORCIDs of your co-authors for the paper, wherever possible
- [ ] In your paper for `... the rapidTide python library ...` capitalize `Python`
- [ ] In your paper for `...  in healthy subjects it is even more true for stroke patients ...`, please add a comma here: `... healthy subjects, it is ...`
- [ ] In your paper for `... However, fMRI signal is inherently noisy ...`, please change to either `... fMRI signals are inherently ...` or to `... the fMRI signal is inherently ...`
