+++ paper/paper.md
@@ -11,16 +11,16 @@ authors:
     orcid: 0009-0008-2046-3703
     affiliation: "1, 2" 
   - name: Giulia Lioi
-    orcid: 
+    orcid: 0000-0002-8327-9814
     affiliation: "1" 
   - name: Julie Coloigner
-    orcid: 
+    orcid: 0000-0001-7774-8015
     affiliation: "2" 
   - name: Pierre Maurel
-    orcid: 
+    orcid: 0000-0003-2539-7414
     affiliation: "2"
   - name: Nicolas Farrugia
-    orcid: 
+    orcid: 0000-0001-7160-0431
     affiliation: "1"
    
 
@@ -34,14 +34,14 @@ bibliography: paper.bib
 ---
 
 # Summary
-Functional Magnetic Resonance Imaging (fMRI) is a widely used neuroimaging technique for the analysis of neural activity and functional connectivity. However, fMRI signal is inherently noisy and susceptible to various artifacts, compromising the accuracy and reliability of derived analyses. This becomes particularly critical when dealing with stroke patients, given the added complexitiy associated with their neurological condition. Specific preprocessing and denoising are integral steps to identify the nuisance sources and mitigate their effect on fMRI analysis. 
+Functional Magnetic Resonance Imaging (fMRI) is a widely used neuroimaging technique for the analysis of neural activity and functional connectivity. However, the fMRI signal is inherently noisy and susceptible to various artifacts, compromising the accuracy and reliability of derived analyses. This becomes particularly critical when dealing with stroke patients, given the added complexitiy associated with their neurological condition. Specific preprocessing and denoising are integral steps to identify the nuisance sources and mitigate their effect on fMRI analysis. 
 To address these challenges, we present fMRIStroke, a dedicated preprocessing pipeline designed specifically for the quality control and preprocessing of fMRI data from stroke patients. fMRIStroke operates as an enhancement to standard preprocessing workflows. Building on the outputs from commonly used tools like fMRIprep, fMRIStroke introduces additional quality control visualizations, computes supplementary confounding variables, and performs confound regression (denoising), resulting in preprocessed fMRI data that is ready for subsequent analysis of neural activity or connectivity.
 
 
 # Statement of need
 Stroke, with its complex and varied impact on the brain's vascular supply, often results in damages that extend beyond the immediate lesion site, making functional connectivity analysis a promising tool to understand its immediate impact as well as network reorganizations during recovery. 
 
-However, while the Blood Oxygen Level Dependent (BOLD) signal obtained from fMRI data is typically mixed with non-neuronal sources of variability [@Bianciardi:2009] in healthy subjects it is even more true for stroke patients [@Siegel:2017], compromising the reliability of fMRI analyses and specifically Functional Connectivity (FC) analysis. 
+However, while the Blood Oxygen Level Dependent (BOLD) signal obtained from fMRI data is typically mixed with non-neuronal sources of variability [@Bianciardi:2009] in healthy subjects, it is even more true for stroke patients [@Siegel:2017], compromising the reliability of fMRI analyses and specifically Functional Connectivity (FC) analysis. 
 Recognizing these challenges, it is highly recommended, notably by @Siegel:2017, to incorporate specific quality controls and strategies to address lesion-related confounds when working with stroke fMRI data. 
 
 Existing tools like fMRIprep [@Esteban:2019] have streamlined most of the preprocessing steps; however, they lack specificity for fMRI data from stroke patients. 
@@ -49,7 +49,7 @@ Existing tools like fMRIprep [@Esteban:2019] have streamlined most of the prepro
[…]
