### 2024-11-18 #6636 @sappelhoff
great, thanks a lot. However, I now see two archive releases on Zenodo:

![image](https://github.com/user-attachments/assets/ce2617d2-6258-4009-94f6-7f68272d4715)

could you please delete the earlier one with the DOI https://doi.org/10.5281/zenodo.14175275 ? to avoid confusion.

Other than that I am fully satisfied and happy to recommend this project for acceptance.
