### 2024-11-18 #6636 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
