### 2024-11-19 #6636 @Kevin-Mattheus-Moerman
@alixlam as AEiC for JOSS I will now help to process this submission for acceptance in JOSS. Below are some final checks I have been working on, some of which require your attention:

- [ ] Please edit the archive listed license to be `v1.0.0`, i.e. to include the `v`
