### 2024-11-19 #6636 @sappelhoff
@Kevin-Mattheus-Moerman just out of curiosity, do you have a reasoning to pick the version with `v` as a prefix? rather than `1.0.0`
