### 2024-11-19 b7f425e
Update paper.md
