### 2024-11-19 #6636 @Kevin-Mattheus-Moerman
> @Kevin-Mattheus-Moerman just out of curiosity, do you have a reasoning to pick the version with `v` as a prefix? rather than `1.0.0`

@sappelhoff we need it here because the tag they used to create a release on their GitHub repository has the v. It would not be needed if there chosen tag just read 1.0.0.
