### 2024-11-20 #6636 @Kevin-Mattheus-Moerman
@alixlam Thanks for making those changes. I think all looks good now. Below is one minor final note.

- [ ] I just noticed that the last sentence contains `... in more details`, that should probably read `...in more detail`.
